⚡️ Optimize SafeCastLib again (#933)
* ⚡️ Optimize SafeCastLib again (#929)

* Add `testSafeCastUint256ToIntBench` function and update .gas-snapshot

A new function, testSafeCastUint256ToIntBench, was added for testing uint256 to int conversion. Also, .gas-snapshot file was updated to reflect changes in gas usage.

* Optimize `SafeCastLib` further

The check for overflow in the SafeCastLib library's typecasting functions was refactored. The conditions for triggering overflow and typecasting were reversed to improve code readability and understanding. The updated statements are also more efficient by avoiding unnecessary shift operations.

* Optimize runtime gas for large optimizer runs

* Refactor and optimize `toInt256`

* Refactor and optimize `toUint256`

* Rename test bench functions

* Add comment on optimization

* Add halmos equivalence test

---------

Co-authored-by: Shuhui Luo <107524008+shuhuiluo@users.noreply.github.com>